model: read V29 marksurface and face indices as unsigned (uhexen2-uwam)
Mod_LoadLeafs_V29 read firstmarksurface and nummarksurfaces as signed
short, so a map with more than 32K marksurfaces wrapped negative and
pointed before the array, reading and writing out of bounds at load time.
The marksurfaces lump loader itself already reads unsigned short, so the
data admits up to 65535 either way.

Mod_LoadNodes_V29 had the same pattern on firstface/numfaces, and the
faces lump is likewise counted as an int, so cast both. Node children
stay signed: negative values there encode leaf indices.

+++ engine/h2shared/gl_model.c
@@ -1548,8 +1548,8 @@ static void Mod_LoadNodes_V29 (lump_t *l)
 		p = LittleLong(in->planenum);
 		out->plane = loadmodel->planes + p;
 
-		out->firstsurface = LittleShort (in->firstface);
-		out->numsurfaces = LittleShort (in->numfaces);
+		out->firstsurface = (unsigned short)LittleShort (in->firstface);
+		out->numsurfaces = (unsigned short)LittleShort (in->numfaces);
 
 		for (j = 0; j < 2; j++)
 		{
@@ -1675,8 +1675,11 @@ static void Mod_LoadLeafs_V29 (lump_t *l)
 		p = LittleLong(in->contents);
 		out->contents = p;
 
-		out->firstmarksurface = loadmodel->marksurfaces + LittleShort(in->firstmarksurface);
-		out->nummarksurfaces = LittleShort(in->nummarksurfaces);
+		/* Unsigned: the marksurfaces lump itself is read as unsigned
+		 * short, so a map with >32K marksurfaces would otherwise wrap
+		 * negative and point before the array. */
+		out->firstmarksurface = loadmodel->marksurfaces + (unsigned short)LittleShort(in->firstmarksurface);
+		out->nummarksurfaces = (unsigned short)LittleShort(in->nummarksurfaces);
 
 		p = LittleLong(in->visofs);
 		if (p == -1)
